switch default from float to double

+++ src/diff.h
@@ -9,7 +9,7 @@
 // root diff namespace segment
 namespace diff {
 
-typedef float constant;
+typedef double constant;
 
 };
